update package & 剩余埋点&logou 回调

+++ Assets/MirrorMarketSDK/Runtime/Services/Interfaces/IMarketplaceService.cs
@@ -33,7 +33,7 @@ namespace MirrorworldSDK.Interfaces
         public void CancelNFTListing(string mintAddress, float price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
 
         //buy
-        public void BuyNFT(string mintAddress, decimal price, Action<CommonResponse<ListingResponse>> callBack);
+        public void BuyNFT(string mintAddress, float price, Action<CommonResponse<ListingResponse>> callBack);
 
         //transfer
         public void TransferNFT(string mintAddress, string walletAddress, Action<CommonResponse<ListingResponse>> callBack);
